helper y filtro admin

+++ blog/app/Http/Controllers/AdminController.php
@@ -5,22 +5,37 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Producto;
 use App\User;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Controllers\Helper;
+
+//Agregar Helper::esAdmin a todos los metodos para validacion de administradores
 
 class AdminController extends Controller
 {
+    
     public function index()
     {
+        if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
+        
         return view('admin'); 
     }
 
     public function productos()
     {   
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $productos = Producto::all();
         return view('productos-admin', compact ('productos'));
     }
 
     public function mostrarAgregar()
     {
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         return view('agregarProductos-admin');
     }
 
@@ -28,7 +43,9 @@ class AdminController extends Controller
 
     public function agregarProducto(Request $req)
     {
-        
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $producto = new Producto();
 
         $ruta = $req->file("imgProducto")->store("public");
@@ -49,6 +66,9 @@ class AdminController extends Controller
     }
 
     public function formularioModificarProducto($id){
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $producto= Producto::find($id);
 
         return view('modificar', compact('producto'));
@@ -56,6 +76,9 @@ class AdminController extends Controller
 
     public function modificarProducto(Request $req)
     {
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         //Traigo el producto que quiero modificar
         $id = $req["id"];
         
@@ -89,6 +112,9 @@ class AdminController extends Controller
 
 
     public function eliminarProducto($id){
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $post = Producto::find($id);
         $post->delete();
     
@@ -98,6 +124,9 @@ class AdminController extends Controller
 
 
     public function clientes(){
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $clientes = User::all();
 
 
@@ -105,6 +134,9 @@ class AdminController extends Controller
     }
 
     public function clientesEliminar($id){
+         if (Helper::noEsAdmin()){
+            return redirect("/");
+        }
         $cliente = User::find($id);
